+++ packages/moonstone/MoonstoneDecorator/MoonstoneDecorator.js
@@ -83,7 +83,18 @@ const MoonstoneDecorator = hoc(defaultConfig, (config, Wrapped) => {
 	addAll({
 		cancel: 461,
 		pointerHide: 1537,
-		pointerShow: 1536
+		pointerShow: 1536,
+		pointerKeys: [
+			461,
+			415, // play
+			19, // pause
+			403, // red
+			404, // green
+			405, // yellow
+			406, // blue
+			33, // channel up
+			34 // channel down
+		]
 	});
 
 	const Decorator = class extends React.Component {